+++ project/OsEngine/OsTrader/Panels/PanelsGui/SmaStochasticUi.xaml.cs
@@ -43,7 +43,7 @@ namespace OsEngine.OsTrader.Panels.PanelsGui
             try
             {
 
-                if (Convert.ToInt32(TextBoxVolumeOne.Text) <= 0 ||
+                if (Convert.ToDecimal(TextBoxVolumeOne.Text) <= 0 ||
                     Convert.ToInt32(StochUp.Text) <= 0 ||
                     Convert.ToInt32(StochDown.Text) <= 0 ||
                     Convert.ToInt32(Step.Text) <= 0 ||
@@ -58,7 +58,7 @@ namespace OsEngine.OsTrader.Panels.PanelsGui
                 return;
             }
 
-            _strategy.VolumeFix = Convert.ToInt32(TextBoxVolumeOne.Text);
+            _strategy.VolumeFix = Convert.ToDecimal(TextBoxVolumeOne.Text);
             _strategy.Slipage = Convert.ToDecimal(TextBoxSlipage.Text);
 
             _strategy.Upline = Convert.ToDecimal(StochUp.Text);
