+++ project/OsEngine/OsTrader/Panels/PanelsGui/ThreeSoldierUi.xaml.cs
@@ -40,7 +40,7 @@ namespace OsEngine.OsTrader.Panels.PanelsGui
             try
             {
 
-                if (Convert.ToInt32(TextBoxVolumeOne.Text) <= 0 ||
+                if (Convert.ToDecimal(TextBoxVolumeOne.Text) <= 0 ||
                     Convert.ToDecimal(TextBoxSlipage.Text) <= 0 ||
                     Convert.ToInt32(TextBoxheightSoldiers.Text) <= 0 ||
                     Convert.ToInt32(TextBoxprocHeightTake.Text) <= 0 ||
@@ -56,7 +56,7 @@ namespace OsEngine.OsTrader.Panels.PanelsGui
             }
             Enum.TryParse(ComboBoxRegime.Text, true, out _strategy.Regime);
             _strategy.Slipage = Convert.ToDecimal(TextBoxSlipage.Text);
-            _strategy.VolumeFix = Convert.ToInt32(TextBoxVolumeOne.Text);
+            _strategy.VolumeFix = Convert.ToDecimal(TextBoxVolumeOne.Text);
             _strategy.HeightSoldiers = Convert.ToInt32(TextBoxheightSoldiers.Text);
             _strategy.MinHeightSoldier = Convert.ToInt32(TextBoxminHeightSoldier.Text);
             _strategy.ProcHeightTake = Convert.ToInt32(TextBoxprocHeightTake.Text);
